Directly link to support instead of referring to Quick Links
In my opinion, it is better to directly link to the support instead of saying to click the link below.
This is more convenient as you don't have to look for the link further down.

Besides that, it is also easy for unexperienced users to click the "Customer Support" link instead of the "Community Support" or vice versa. Directly linking to the correct support mitigates that.