Add a message for reports duplicating MCL-14369 (#114)
* Add MCL-14369 helper message

* I forgot to change the name after copy-and-paste

Oops =P

* Finalized the draft of the MCL-14369 message

I changed the shortcut to 'mve' (missing version/configuration). Feel free to change or request a change if this is not ideal.

* Make consistent with #116

Fixes the text to be consistent with #116.

* Modify MCL-5638 helper message

More specifically, changed "Moderator's Note" to "Moderator Note". This makes it consistent with the MCL-14369 message and with the panel title from that report.