Hide not-reopen-ar message
The message tells the user that a moderator will update the report, or that the user should
contact staff. Therefore it does not make sense that a staff member would use this message
because they would directly interact with the user, e.g. ask them for the information.

This message only makes sense for the bot Arisa.