Make message names more consistent
Most of the messages using imperative mood describe what the user should
do, however "Panel - Mark as private" is about what the staff member / Arisa
did, therefore imperative mood should not be used there. Using past tense
instead might be irritating because most message names are about what the
user did / should do. Therefore naming it "Marked as private" could be
misundertood as something which the user did, however the action was
performed by staff / Arisa.

This is similar for "Panel - Unmark private issue" as well. It describes an action
the user did in the past and therefore it is more appropriate to use the past
tense.